refactor: making revision come back to portifolio
